[AMDGPU] Improve abs modifier usage
If a call to the llvm.fabs intrinsic has users in another reachable
BB, SelectionDAG will not apply the abs modifier to these users and
instead generate a v_and ..., 0x7fffffff instruction.
The same logic applies for the fneg instruction.
This patch implements shouldSinkOperands for the AMDPGPU target,
so CodeGenPrepare can try sinking / cloning the operands to the
basic blocks of the users.

Differential Revision: https://reviews.llvm.org/D150347