Add --split-same-checks option to `update_test_checks.py`
With very large argument lists and metadata substitutions auto-generated
in lit tests, we can run into an issue where `FileCheck` tries to
generate back-references with a look-ahead larger than `9`, which is
currently not supported. A workaround here is to split the `-SAME` check
lines into multiple check lines such that the back-reference regex gets
a smaller search space.